refactor: ChallengeCard Component가 분산되어있는 이슈 수정
