+++ src/types/challenge.ts
@@ -1,0 +1,9 @@
+export type ChallengeType = {
+  id: string;
+  thumbnail: string;
+  title: string;
+  person: number;
+  point: number;
+  tags: string[];
+  isProved?: boolean;
+};
